Expose the servicing of l1 cache misses cache through the wrapper interface
The original Spike cache considered the servicing of L1 misses as instantaneous and did not wait
for the L2. A new class representing a serviceable cache has been added. This is identical to
the base Spike cache, but the handling of replacement has been extracted to an independent function
that is exposed through the interface. This function will be called once a L2 request has been serviced.
This function might generate a writeback request to the L2.