Reduce the number of generated MCPUSetVVL packets
The sent MCPUSetVVL packets are to inform the MemTile about a new LMUL, SEW or AVL setting. However, the vsetvl instruction can also be used to update neither of this information. Specifically in the following cases, no MCPUSetVVL packet needs to be sent:
1) RVV 0.8, p. 26 states: "When rs1=x0 and rd=x0, the current vector length in vl is used as the AVL, and the resulting value is only written to vl."
This is an internal setting that affects the VPU, but not the MemTile. VVL remains unchanged.
2) If there is no change in the AVL nor in the vtype register (containing LMUL and SEW), then there is no need to send a packet.