Removing vsetvli and vector memory operation dependencies.
For memory operations, the MCPU will ask for VVL elements from the scratchpad (vstore) or transfer VVL elements to the SP (vload). However, the VAS Tile does not need to know the VVL setting for either of those memory operations. Hence, the RAW dependency has been removed in this commit.