Fixed a bug, in which the new VL setting was not saved
In case, that there is no need to inform the MemTile about a VVL changed (e.g., if only the vtype register changes), the destination register given in the vsetvli instruction was not updated. This commit fixes this issue. In addition, per definition, if rs1=rd=0, the requested VL needs to be (uint_t)-1.