Reverting dependency of mem_ops on vsetvli, dependency as macro
In commit da879593c9e540db9d06a9cb868ebb2d200beebf, this dependency was partially removed. However, while it is true that for a loading memory operation, VVL is not necessary, it is necessary for following vector arithmetics. There to fire the computation, the vector operands need to be available (e.g. through a vector load) and VVL needs to be current as well. There is no dependency implemented between the compute instruction and the vsetvli. Therefore, to get this dependency, the following dependencies are followed for correctness: vsetvli -> vector_mem_op -> vector_compute. Therefore, the vsetvli -> vector_mem_op dependency has been reinstated.