vlmax is computed by the memory tile.
If vsetvli is executed and the maximum VL is requested, the MemTile computes this VL and returns it. That means, in case the MemTile is enabled, the core should take this value from the MemTile (which can be lesser than what is physically possible due to some constraints or if the MemTile decides, that L2 space is more important than SP space) and not the vlmax value that the VAS Tile computed when starting up.